backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pandas==2.1.1
nltk==3.8.1
redis==4.6.0
python-dotenv==1.0.0
websocket-client==1.6.0
feedparser==6.0.11
psycopg2-binary==2.9.11
fakeredis==2.21.0
pytest==7.4.0
transformers==5.2.0
torch==2.10.0
scipy==1.16.3
datasets==4.8.3
xgboost==2.1.4
scikit-learn==1.5.2
shap==0.48.0
APScheduler==3.10.4
